test: add some tests for the pkg: import feature
