+++ fixtures/e2e/pkg-import/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "pkg-import",
+  "version": "1.0.0",
+  "private": true,
+  "type": "commonjs",
+  "description": "Test using the pgk import from a module",
+  "scripts": {
+    "test": "node --test sass.test.js"
+  },
+  "devDependencies": {
+    "sass": "^1.71.0"
+  }
+}
