+++ fixtures/e2e/pkg-import/src/scoped.scss
