+++ fixtures/e2e/pkg-import/src/styles.scss
